# openapi.yaml
openapi: 3.1.0
info:
  title: Monday Secretary API
  version: 1.1.3
servers:
  - url: https://health-api-server.onrender.com
paths:
  /functions/get_health_data:
    post:
      operationId: getHealthData
      summary: 体調データを取得
      tags:
        - Health
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /functions/get_work_data:
    post:
      operationId: getWorkData
      summary: 業務メモを取得
      tags:
        - Work
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetWorkDataRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDataResponse"
  /functions/calendar_event:
    post:
      operationId: calendarEvent
      summary: カレンダー操作
      tags:
        - Calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarResponse"
  /memory:
    post:
      operationId: createMemory
      summary: Notion にメモを保存
      tags:
        - Memory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryRequest"
            example:
              title: "眠い"
              summary: "会議中にウトウトした"
      responses:
        "200":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryResponse"
  /functions/confirm_memory:
    post:
      operationId: confirmMemory
      summary: ユーザーが記憶保存を承認した時に呼ぶ
      tags:
        - Memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
              properties:
                session_id:
                  type: string
      responses:
        "200":
          description: OK
  /functions/get_memory:
    post:
      operationId: getMemory
      summary: 記憶（Notion ページ）を検索
      tags:
        - Memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer
  schemas:
    HealthRequest:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - latest
            - compare
            - period
            - dailySummary
          default: latest
        sheet_url:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    HealthItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        sleep_hours:
          type: number
        slept_well:
          type: boolean
        weight:
          type: number
        energy:
          type: string
        mood:
          type: string
        memo:
          type: string
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/HealthItem"
    GetWorkDataRequest:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - latest
            - period
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    WorkItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        box_setup_9am:
          type: boolean
        tasks_after_10am:
          type: string
        rest_minutes_am:
          type: number
        fatigue_level_am:
          type: string
        comment_am:
          type: string
        box_collected_3pm:
          type: boolean
        tasks_pm:
          type: string
        rest_minutes_pm:
          type: number
        fatigue_level_pm:
          type: string
        comment_pm:
          type: string
        daily_summary:
          type: string
    WorkDataResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: "#/components/schemas/WorkItem"
    CalendarRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            - insert
            - get
            - update
            - delete
        calendar_id:
          type: string
        summary:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        event_id:
          type: string
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        summary:
          type: string
        calendar_id:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    CalendarResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        event:
          $ref: "#/components/schemas/CalendarEvent"
    MemoryRequest:
      type: object
      required:
        - title
        - summary
      properties:
        title:
          type: string
        summary:
          type: string
        category:
          type: string
          default: ""
        emotion:
          type: string
          default: ""
        reason:
          type: string
          default: ""
        timestamp:
          type: string
          format: date-time
          description: ISO 形式日時。省略時は現在時刻
    SavedMemory:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        created_time:
          type: string
          format: date-time
    MemoryResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        saved_memory:
          $ref: "#/components/schemas/SavedMemory"
    MemorySearchRequest:
      type: object
      properties:
        query:
          type: string
          default: ""
        top_k:
          type: integer
          default: 5